requirements.txt
pandas
numpy
python-dotenv # to read the .env file
mysql-connector-python
pymysql 
scikit-learn
#-e